Refactoring Phase 2
- Moved tickets into structres dir.
- Created new dir constants
- Renamed constants.rs into tss.rs and moved into constants dir.
- Moved algorithm/specifiers.rs into constants dir and renamed
  it to algorithm.rs.
- Created tags as subdir to constants.
- Renamed utils/tags.rs to structure.rs and moved it into the
  constants/tags dir.
- Moved error part of respopnse_code.rs into sepparate file error.rs
  and moved response_code.rs into constants dir.
- Changed lints in lib.rs to make more sense.
- Moved lint from lib.rsd to constants/mod.rs
