+++ src/algorithm/mod.rs
@@ -1,4 +1,3 @@
 // Copyright 2020 Contributors to the Parsec project.
 // SPDX-License-Identifier: Apache-2.0
-pub mod specifiers;
 pub mod structures;
