+++ src/algorithm/structures/sensitive/data.rs
@@ -1,7 +1,7 @@
 // Copyright 2020 Contributors to the Parsec project.
 // SPDX-License-Identifier: Apache-2.0
-use crate::response_code::{Error, Result, WrapperErrorKind};
 use crate::tss2_esys::TPM2B_SENSITIVE_DATA;
+use crate::{Error, Result, WrapperErrorKind};
 use log::error;
 use std::convert::TryFrom;
 /// Rust native representation of sensitive data.
