+++ src/macros/mod.rs
@@ -1,0 +1,2 @@
+// Copyright 2020 Contributors to the Parsec project.
+// SPDX-License-Identifier: Apache-2.0
