+++ src/structures/lists/pcr_selection.rs
@@ -1,10 +1,9 @@
 // Copyright 2020 Contributors to the Parsec project.
 // SPDX-License-Identifier: Apache-2.0
-use crate::algorithm::specifiers::HashingAlgorithm;
-use crate::response_code::{Error, Result, WrapperErrorKind};
+use crate::constants::algorithm::HashingAlgorithm;
 use crate::structures::{PcrSelectSize, PcrSelection, PcrSlot};
 use crate::tss2_esys::TPML_PCR_SELECTION;
-// use enumflags2::BitFlags;
+use crate::{Error, Result, WrapperErrorKind};
 use log::error;
 use std::collections::HashMap;
 use std::convert::TryFrom;
@@ -54,7 +53,7 @@ impl PcrSelectionList {
     /// # Examples
     /// ```
     /// use tss_esapi::structures::{PcrSelectionListBuilder, PcrSlot};
-    /// use tss_esapi::algorithm::specifiers::HashingAlgorithm;
+    /// use tss_esapi::constants::algorithm::HashingAlgorithm;
     /// // pcr selections
     /// let mut pcr_selection_list = PcrSelectionListBuilder::new()
     ///     .with_size_of_select(Default::default())
