+++ tests/pcr_selection_list_builder_tests.rs
@@ -2,8 +2,8 @@
 // SPDX-License-Identifier: Apache-2.0
 use num_traits::ToPrimitive;
 use std::convert::TryFrom;
-use tss_esapi::algorithm::specifiers::HashingAlgorithm;
-use tss_esapi::constants::TPM2_ALG_LAST;
+use tss_esapi::constants::algorithm::HashingAlgorithm;
+use tss_esapi::constants::tss::TPM2_ALG_LAST;
 use tss_esapi::structures::{PcrSelectSize, PcrSelectionList, PcrSelectionListBuilder, PcrSlot};
 use tss_esapi::tss2_esys::{TPM2_ALG_ID, TPML_PCR_SELECTION};
