+++ tests/pcr_selection_list_tests.rs
@@ -1,6 +1,6 @@
 // Copyright 2020 Contributors to the Parsec project.
 // SPDX-License-Identifier: Apache-2.0
-use tss_esapi::algorithm::specifiers::HashingAlgorithm;
+use tss_esapi::constants::algorithm::HashingAlgorithm;
 use tss_esapi::structures::{PcrSelectSize, PcrSelectionListBuilder, PcrSlot};
 
 mod test_pcr_selection_list {
